[sw] Move check.h to sw/device/lib/testing
This makes it clearer that the `CHECK` macro is intended only for
non-production code.

This commit also removes an unneeded `#include` in check.h, and updates
the includes in all files that use the CHECK macro.
