+++ sw/device/benchmarks/coremark/top_earlgrey/core_portme.c
@@ -10,7 +10,7 @@
 #include "sw/device/lib/arch/device.h"
 #include "sw/device/lib/base/stdasm.h"
 #include "sw/device/lib/dif/dif_uart.h"
-#include "sw/device/lib/runtime/check.h"
+#include "sw/device/lib/testing/check.h"
 #include "sw/device/lib/testing/test_status.h"
 #include "sw/vendor/eembc_coremark/coremark.h"
