+++ sw/device/tests/dif/dif_rv_timer_sanitytest.c
@@ -6,10 +6,10 @@
 
 #include "sw/device/lib/handler.h"
 #include "sw/device/lib/irq.h"
-#include "sw/device/lib/runtime/check.h"
 #include "sw/device/lib/runtime/hart.h"
 #include "sw/device/lib/runtime/ibex.h"
 #include "sw/device/lib/runtime/log.h"
+#include "sw/device/lib/testing/check.h"
 #include "sw/device/lib/testing/test_main.h"
 #include "hw/top_earlgrey/sw/autogen/top_earlgrey.h"
