+++ sw/device/tests/pmp_sanitytest_tor.c
@@ -6,8 +6,8 @@
 
 #include "sw/device/lib/handler.h"
 #include "sw/device/lib/irq.h"
-#include "sw/device/lib/runtime/check.h"
 #include "sw/device/lib/runtime/hart.h"
+#include "sw/device/lib/testing/check.h"
 #include "sw/device/lib/testing/test_main.h"
 #include "sw/device/lib/testing/test_status.h"
